+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "lint": "standard",
     "lint:fix": "standard --fix",
-    "dev": "nodemon zowi.js", 
+    "dev": "nodemon zowi.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -24,8 +24,10 @@
     "standard": "^12.0.1"
   },
   "dependencies": {
+    "@serialport/parser-readline": "^8.0.4",
     "express": "^4.17.1",
     "node-bluetooth": "^1.2.5",
+    "serialport": "^8.0.5",
     "socket.io": "^2.2.0"
   }
 }
